backend/agents/research_agent.py
import requests
from bs4 import BeautifulSoup
from typing import Dict, Any, List
import json
import os

class ResearchAgent:
    def __init__(self, serpapi_key: str = None):
        self.serpapi_key = serpapi_key

    async def search_stats(self, location: str, incident_type: str) -> Dict[str, Any]:
        """
        Search for road safety stats using SerpAPI or Scraper.
        """
        if self.serpapi_key:
            # Placeholder for actual SerpAPI call
            # Example query: "Road safety statistics for {location} {incident_type}"
            pass
            
        # Return mock statistical data for demo
        return {
            "region": location,
            "annual_accidents": 5234,
            "annual_deaths": 1567,
            "common_causes": ["Overspeeding: 42%", "Drunk driving: 28%", "Distracted: 18%"],
            "trend": "Decreasing by 5% YoY",
            "sources": ["morth.nic.in", "who.int", "delhitrafficpolice.nic.in"]
        }

    async def search_guidelines(self, incident_type: str) -> List[str]:
        """
        Fetch government guidelines.
        """
        return [
            "Speed limit: 80 km/h on highways",
            "Helmet mandatory for two-wheelers",
            "Seatbelt mandatory for all passengers",
            "Follow 3-second gap rule strictly"
        ]

    async def search_prevention(self, cause_keywords: List[str]) -> List[str]:
        """
        Fetch prevention tips. 
        """
        tips = [
            "Maintain safe distance at all times",
            "Avoid night driving on highways unless necessary",
            "Regularly check tyre pressure and brake efficiency",
            "Avoid sudden lane changes without signaling"
        ]
        return tips

    async def search_authorities(self, location: str) -> Dict[str, Any]:
        """
        Find local authority contact information.
        """
        return {
            "police": "100",
            "ambulance": "108",
            "fire": "101",
            "local_rto": "011-12345678",
            "nearest_hospital": "AIIMS Trauma Center - 2km"
        }
