backend/requirements.txt
streamlit
pdfplumber
pytesseract
pdf2image
pillow
reportlab
python-docx
python-pptx
matplotlib
openai
google-generativeai
serpapi
requests
python-dotenv
msal
beautifulsoup4
pandas
numpy
fastapi
uvicorn
python-multipart
